213: error handling for CRN search
We want to return error messages when users search for a CRN. This checks client side if a CRN has been entered adds a flash message to be rendered by the applications controller if it has not. This also handles the possibility of a 404 or 403 from the server, and throws any other error.