218: Setup Sentry request and error handling
We will be using Sentry to track errors in the app. We're following the
same implementation as CAS1 + 2[1], with slight difference that we do not
need to use the `@sentry/tracing` package because it is being
deprecated.[2]

We are also removing the username from the user sent to Sentry (a la
Accredited Programmes[3]) because it was still being logged despite
anonymising the data.

[1] ministryofjustice/hmpps-temporary-accommodation-ui@f4fd6ce
[2] https://docs.sentry.io/product/sentry-basics/tracing/
[3] ministryofjustice/hmpps-accredited-programmes-ui@c645f1a