218: test and debug sentry
in order to have confidence that sentry errors are being logged, we need
to create errors for the environment to be seen in Sentry [1] This debug
test allows those errors to be created. I've also moved the integration
tests for login and health out of the legacy `e2e` folder where they were not
being picked up by cypress.

[1]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/commit/2acadacbbab80a498c6675b3de905011a4b7d6d1