add status type
we would like to be able to render applications a user has
created, and indicate the ones that are still in progress.

this commit pulls in the updated `Cas2ApplicationSummary` model from the
API types, with the new `status` field, which we will use to determine
if an application is in progress.