add GroupedApplications type
we would like to be able to render applications a user has
created, and indicate the ones that are still in progress.

This commit adds an interface to show the shape of applications that
have been grouped by their status. Currently the only available status
is `inProgress` but there will be `submitted` apps in the near future.