return in progress applications for a user
we would like to be able to render applications a user has
created, and indicate the ones that are still in progress.

This commit renames the `applicationService` function to be more
accurate, and adds logic to group applications by their status. (Copied
over from CAS1/3). We only have in progress apps at the moment but
submitted ones will come in the future.

This relies on the API work having been done here to filter applications
by user id.
https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/840