add `auditService`
We want to integrate with HMPPS audit service and track our users
actions in the service.

This commit adds CAS3's audit service which we will shortly use in some
middleware.[1]

Modelled after the AuditService in the `make-recall-decision-ui`
project, we add an AuditService for adding messages to a configured AWS
SQS queue. We are able to pass a username through to this service after
discussions with the Audit team [2]

[1]https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/commit/42717adcc2981ce6c9cb8bb74169a8507b61ae1b
[2] https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/108/commits/e82a27de95d86f8ab57421d3b45efba84145c047