add `auditMiddleware`
We want to integrate with HMPPS audit service and track our users
actions in the service.

This commit adds CAS3's audit middleware which we will shortly use to
audit our routes [1]

We add a function that takes a request handler, and returns an audited
version of that request handler. When the audited request handler is
invoked, it will invoke the original request handler, and additionally
send an audit message using our AuditService

[1] https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/pull/101