Add 'Before you start' section with 'Confirm eligibility' task
Initially we add this as the first section in the task list.
In the next commit we will change the flow so that this
comes after searching for CRN, before the task list is seen
for the first time.

We'll then need to rejig the `application.data` state of all
the integration tests so that tasks have the expected
status e.g. 'Not started'...