Move 'Confirm eligibility' task to after application creation
We verify that after successfully searching for a person by
CRN, we are redirected to the "Confirm eligibility" task,
the first task in the "Before you start" section.

`ApplicationsController.show()` now redirects to the
"Confirm eligibility" task unless that question has been
answered.

In the next commit we will tweak the other integration tests
so that they have the new `application.data` state for their
tasks to have the expected status.