Update Equalities questions page and tests
Now that we are developing a subsequent page in the task, I have updated the page and tests to reflect the fact that on success we should be redirected to the disability page.

I have also updated the integration test file name to reflect the face that we are testing a page rather than a task.