Add util functions to submit application
We would like users to be able to mark applications as submitted.

This commit adds the barebones utility function used by CAS1/3 to
transform application data for submission - the `translatedDocument`
which will be used to render a completed application.

This is a first pass at the submission flow, and we may need to return
to this functionality when we have more of an understanding of what the
post-submission requirements are.