Add form to Task list
We would like users to be able to mark an application as submitted.

This commit wraps the Task List in a form component to POST to the
submission route. We will not require the confirmation checkbox to be
selected for submission at this time, but leaving it in there for future
iteration.