Implement 'Substance misuse' response()
The `response()` function of each page is a critical representation of the state of
the questions and answers of that page:

- the questions are captured in the current form or 'translation' as copy may be
tweaked from time to time and it's important to record the precise question which
was posed

- answered should be 'translated' as needed e.g. to show 'Any other mixed or multiple
ethnic background' if that was the label shown against the "other" radio value.

For this page's response we put the "yes" value into sentence case.

Application submission
----

In CAS1:

`ApplicationController.submit()` sets the read-only finalised version of the
application (the `document`):

```ts
application.document = getResponses(application)
```

which uses the `getResponses()` utility to collate each page's `response()`:

```ts
getResponses = (formArtifact: FormArtifact): ApplicationOrAssessmentResponse => {
  const responses = {}

  const formSections = getSections(formArtifact)

  formSections.forEach(section => {
    section.tasks.forEach(task => {
      const responsesForTask: Array<PageResponse> = []

      forPagesInTask(formArtifact, task, page => responsesForTask.push(page.response()))
      responses[task.id] = responsesForTask

    })

  })

  return responses
}
```

Check your answers
-----

Building the check your answers page is more involved as the `application.data`
structure doesn't provide all that's needed -- principally the grouping of tasks
into sections, though this may have been added in CAS3?