Add an 'optional' attribute to medicationIssues question
We need to add an 'optional' flag to the `label` element of
questions which aren't mandatory, in this case

"Please list any issues...with medication"

See https://design-system.service.gov.uk/patterns/question-pages/

The string "(optional)" should not be part of the actual
question, it is an attribute or property of the question.

Accordingly, we add an `optional` property to our form page
class.