Verify that '(optional)' flag is shown on question label
We add an `fieldIsOptional` integration test helper so
we can verify that the optional flag has been set as
expected.

Now any text area can have its label given an 'optional'
flag without having to corrupt the question's title.

We will probably soon need this "optional flag" feature to
`formPageInput` also, but I held off from doing this now as
we don't have a use-case with which to verify that it's
working.