add API types and utility functions
If a LAO is returned from the API the UI will now see only their CRN.[1]
This means whenever we retrieve a person or the person on an
application, we need to check whether they are a `FullPerson` type in
order to see their name and other details.

This commit adds the utility functions and API types needed to handle
the changes.

[1] ministryofjustice/hmpps-approved-premises-api#804.