Create risk to self task and guidance page
The guidance page includes an initialize method that will on success return an instance of the class with the relevant OASys data, if data has been returned. If no data has been returned then the class will still be instantiated, and conditional rendered content in the nunjucks template will inform the user that no OASys data has been found.