update form pages to handle LAOs
If a LAO is returned from the API the UI will now see only their CRN.[1]
This means whenever we retrieve a person or the person on an
application, we need to check whether they are a `FullPerson` type in
order to see their name and other details.

This commit changes all instances of looking at the name of a person on
form pages, to use a helper method. If it is not a full person, we use
some placeholder copy.

In theory the user of CAS2 should not be able to create an application
for an LAO so it would be an edge case that they ever see the
placeholder copy, but we need to future proof it in the case that a
prisoner becomes an LAO while application in progress, or in case an
application owner changes to a user who does not have permission to see
LAOs.

[1] ministryofjustice/hmpps-approved-premises-api#804.