update people controller to handle LAOs
If a LAO is returned from the API the UI will now see only their CRN.[1]
This means whenever we retrieve a person or the person on an
application, we need to check whether they are a `FullPerson` type in
order to see their name and other details.

This commit updates the people controller to throw the
`RestrictedPersonError` when an LAO is returned, so that an application
cannot proceed.

[1] ministryofjustice/hmpps-approved-premises-api#804.