refactor Risk To Self page to use new endpoint
CAS2 would like to request only the Risk to Self section of an OASys.
This commit uses a new method to retrive only the Risk To Self section
on the Risk To Self page, and adds a question that was previously not
returned by the API (`historicalRisk`).

We are also recatoring the transformaiton function to use a `switch`
statement, and the assigning of variables to be more clear.