add `documentTitle` to all TaskListPages
We currently use a form page's `title` field to populate the document
title `<title>` element.

In order to protect PII leaking through Google Analytics (and
potentially other platforms) we need to remove the person's name from
this element. We still want to preserve the name in the page heading as
it provides a better user experience.

I'm proposing introducing a new required field on a `TaskListPage` that will
account for what the document title should be, so that we can provide a
custom one without the person's name.

The `<title>` is then set on the shared page `layout` template.

I will shortly introduce an integration test to check on every page that
the person's name is not included.