add `removeFromList` service function
We need to allow users the ability to remove an ACCT from the
application form. The ACCTs will be a list of objects on the `acct-data`
page name in the application data. We will use the index of an acct to
remove it from the list, passed in as a query param from a link on the
`acct` page, which renders the list data.

This adds a function the Application service to remove from a list,
quite similar to the previously added `appendToList`.