add `remove` link to ACCTs
We need to allow users the ability to remove an ACCT from the
application form. The ACCTs will be a list of objects on the `acct-data`
page name in the application data. We will use the index of an acct to
remove it from the list, passed in as a query param from a link on the
`acct` page, which renders the list data.

This commit builds that path with query params for each ACCT.