toggle checkboxes on 'Risk management arrangements'
When 'No arrangements' is checked, any other checkboxes should be
unchecked. When any of the arrangements are checked, then 'No
arrangements' should be unchecked. We should also wipe any data written
in the details input for an arrangement, to avoid the risk of
saving/sharing data that does not need to be.

This commit adds custom javascript to enable this. Unfortunately the ids
for the checkbox elements are not very descriptive. It is possible to
change the ids on the component, but the error message id would then
also have to change (it currently uses the field key `arrangements`)