allow tasks to be completed in any order
Our users would like to be able to dip in and out of the form tasks in
any order. This commit removes the logic that assigns a 'cannot_start`
status to tasks. We are keeping the `cannot_start` status and logic in
the `taskListUtils` because we will in the future be introducing some
rules, e.g. the first task should be completed before the rest.