add integration test for updating application
We currently have some 'custom forms' which save data to other pages or
parts of the application data. There was previously no test coverage to
check that when a form is submitted, it PUTs the data we expect to the
application.

This commit adds a cypress task `verifyApplicationUpdate` to look at the
requests made to check the application update endpoint has been called
with the data we expect. (This is similar to the
`verifyPlacementApplicationSubmit` task on CAS1)

This is a bit of a divergence from our integration tests just testing
what we expect to see on the UI, but I feel it would be valuable to do
on these custom forms, as it will be less expensive than clicking through
to the page we expect to see the data on, and provides coverage not
currently provided by unit tests.