allow for arrangements that are strings or arrays
we would like to be able to have a submitted application flow,
this question was erroring if the user had only checked one 'arrangement', because the checkboxes return an array for multiple answers but just a string for single answers.
This commit adds a quick fix to allow us to submit applications without errors for this page.