move risk summary data to it's own key in application data
The RoSH risk ratings are retrieved from the API and saved to our
application data by the `oasysImport` page. This data is then retrieved
from the application data used to populate the 'risk widget' by the
Summary page.

The risk ratings were previously stored under the key for the Summary
page, `summary`. When the user first lands on the Summary page we are
able to see the widget, but when the user presses 'continue' the form
page takes that `summary` key and uses it to overwrite the data that's
there with it's own form data (see `save` function in the
`applicationService`). Thus deleting the risk widget data.

This commit proposes a solution of saving the risk widget data under its
own key on the application data. This means that it will not be
overwritten by another form. The downside is that it effectively creates
the idea of a `page` in the application data JSON that is not shown to
the user. I think that this is mitigated by the pattern we are already
using for 'add to list' pages, where we use one page to render data and
another page key to store the data (e.g. `acct-data`). This widget is
also a slight anomaly because a) it is not user-inputted data, but an
exact copy from the API  b) The source of truth for application data
should be the JSON itself, not the typescript form pages.