Add check your answers utils
These utils will allow us to generate the cards rendered on the ‘Check your answers’ page. There is some overlap with the code used by CAS1/3 for the same purpose, but they are significantly refactored in line with the ADR ‘Refactor 'review answers' code inherited from CAS1/CAS3’.

For now we are explicitly ignoring RoSH summary and have exceptions in the code for generating ACCT and behaviour notes. These are temporary solutions that we will iterate on shortly.