add custom response to data pages
We would like to decouple as much as possible the Check Your Answers utility function from specific page keys.

For pages that we will not render on CYA:

* we do not need to return anything from the getQuestions util
* we will return an empty object from the response function for the page. This is the purpose of that function, to allow for custom responses rather than specifying in the CYA utilities.

This change removes the need for a hand-rolled list. We still need to include the summary pages currently, but will create a future ticket to render that data in it's own way.