Refactor task answer util for creating document
We want to build a snapshot of the application to save to the
Application model on submission. This requires processing all answers in
the application.

To achieve this we reuse the `getTaskAnswersAsSummaryListItems` function
used for the check your answers page by passing in a
`CheckYourAnswersPage` flag.  This controls whether we return an object
structured for rendering in on the check your answers page (with action
data and html/value keys) or a flat object with the question and answer
for building the document.

In a future refactor, ideally checkYourAnswersUtils will be broken
up to provide this flat object by default and then further transformed
for the CYA page. Given time constraints, I opted to reuse the logic as
is.