render submitted application for assessor
includes an integration test method `hasQuestionsAndAnswersFromDocument`
which is in the base Page class because we will shortly use it to test
the submitted view for an assessor.

I'm not entirely sure about that this way of testing provides much
value, and the tests take a long time to complete and will only get
longer, but I'm also not sure of a better way of doing it.