add custom CSS to funding checkbox
When the divider was passed as the first element to the checkbox
component, the `for` id given to the next checkbox had a `-2` at the
end. This meant that if there was an error summary it was not able to
link to the first checkbox.

I couldn't see a way to override the index-based attributes given to the
checkbox items, so this is a bit of a hacky way of adding a divider
without adding it to the items.