remove funding information conditional data
we have several tasks which have branching logic within them.
E.g. on the funding-source page:
    IF I answer personal savings
    THEN I am taken to national insurance page
    AND I do not see the identification page
    or
    IF I answer benefits
    THEN I am taken to the identification page

If a user returns to the `funding-source` page and changes their answer
from `benefits` to `personalSavings`, then we need to ensure any answers
based on `benefits` are deleted.

Because of the way the page system works, it's tricky to get each page
to automate or handle this logic themselves. We only know what data
needs to be deleted /after/ the page data has been submitted.

I'm hesitant to introduce yet another controller function as we did with
the custom forms[1], because of the boilerplate overhead and the potential
for more complexity. This commit suggests that this data manipulation is
handled by the `save` function in the application service.

This `save` function is called by every standard page, and contains
other logic about how we handle the data. If this change is accepted, we
could go on to add more pages, such as equality and diversity, and
refactor the function to make it easier to add more pages.

[1] doc/architecture/decisions/0002_custom_forms.md