add a timeline template
we want to show a timeline of events for an application on an application overview page.
we are using the raw HTML here because of a problem with just using the component - it's not able to import a certain date module. This is also copying the way that CAS 1 has implemented theirs.