silently handle errors for OASyS import
If there is an error retrieving OASyS data, we do not want to surface
that error to the user. We instead want to indicate that it was not
possible to retrieve an oasys report at this time - the user is then
given the option to proceed with the form and fill it in manually, or
return another time.

This is to quickly fix an issue in development where if the API is not
available, then the user is not able to proceed with form. In the future
we may want to show a more helpful message based on the type of error
received, e.g. 'Person has not been found', 'OASyS API is not available
at this time'. We may want to consider what error to show if the user is
unauthorized as well.

This is in line with how CAS1 handle errors
regarding OASyS [1], catching all instances of errors and allowing the
user to proceed with a blank form.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-ui/blob/main/server/utils/oasysImportUtils.ts#L42