search by prison number
We currently have the functionality to search (and add) a person by CRN.
We received feedback that not all POMs would know how to get this ID,
and the "prison number" (aka NOMIS ID, aka nomsNumber on the API) would
be more appropriate. For the simplicity of the MVP we have decided to
remove searching by CRN and only enable searching by prison number.