Add current offences page
I have created an initialize method to redirect the user to the current offences data page if it is their first time accessing the task.

I also added an ‘offenceList’ body property to allow us to add an error summary to the top of the page if the user attempts to save after removing all offences.