add the `post` route to create an application
We will shortly be adding a 'confirm applicant details' page that will
be shown to the user once they have searched for an offender using a
prison number, allowing them to check they have the right person before
creating an application.

This will mean moving the 'create application' functionality out of the
people controller where it currently sits, as we have been creating an
application as soon as a person is 'found'. Instead the 'confirm
applicant' page will be a form that POSTs the crn to the applications
controller.