Fix check your answers code
Until now, on all the pages where we are relying on a response method, we're pulling the data manually from the application, rather than from the page body, so we haven’t noticed that we aren’t passing the body in as we should be. This fixes that bug, ahead of introducing the CPP details page, which will require it.