Add confirm consent page
In the constructor we remove from the body any data relating to the option that has not been selected, to avoid this data being saved to the database and confusing the user.

I have also added a response method to handle the date data. This method supports a consent refused answer, although it is unclear if/where this would ever be rendered.