Use fieldset for working mobile phone page heading
We want to be able to e2e test this page. As there are two identical
radio inputs with the "Yes" label, we want a way to differentiate
between the two.

We do this by using the page heading as the fieldset `legend`, to
associate the heading with the form.