Hide back link and form on print version of check your answers
We don't want unnecessary elements appearing in the printed version of
the check your answers page so here we remove them.

We keep the confirmation copy in the printed version but remove the form
elements.

We also don't want the hidden elements to continue to take up space on
the page. Changing the CSS property from `visibility: hidden` to
`display: none` ensures these elements are completely removed.