Add phase banner
We need to show a GDS phase banner[1] on all pages of the app in order to capture feedback from users or receive emails from them.

This commit adds the component to the header partial. As with CAS1, we decided to sit it below the `mojPrimaryNavigation` because it did not look quite right sitting between the main header and the nav. Although the GDS guidance says to put it directly under the header, there's some discussion about this being the right place [2] and this seems like the right thing for us.

[1] https://design-system.service.gov.uk/components/phase-banner/
[2] alphagov/govuk-design-system-backlog#57