Add tests for deleteOrphanedData module
Context:

We required a way to remove data that had been given under conditional
questions in the questionnaire. `deleteOrphanedData` was created to
manage this, but has since grown to the point where we need an
alternative.

This new solution will be implemented in follow-up work with the plan
of moving some of the logic from here to the new solution in future.
This change ensures that `deleteOrphanedData` is tested for the pages
currently implemented and will give us confidence when we refactor this
logic in the future.