Add a custom response to the current risk page
This allows us to render the OASys dates in check your answers and the submitted application view. If there is data under the oasys-import key, the OASys dates are returned along with the current risk information.

As it stands this will render the OASys dates at the top of the risk to self section in check your answers. It is not an ideal solution because it means the current risk page response method handling data unrelated to current risk, and if the order in which data is saved to the application were to change, these dates would no longer appear at the top of that section.

For now we have decided that this is a satisfactory solution, but if we encounter similar cases in the future, this would be a good candidate for refactoring.