Rename service name so it matches the production URL
When a user clicks a link `short-term-accommodation-cas-2` then they should feel confident that they arrived on the right place so we make the URL and name consistent.

The start page still references “Community Accommodation Service Tier 2” in the page copy, we leave that as it is in the knowledge that this page is just about to have a larger change.