Add optional onSave method to TaskListPage
We required a way to remove data that had been given under conditional
questions in the questionnaire. `deleteOrphanedData` was created to
manage this, but has since grown to the point where we need an
alternative.

We add the ability for a page to make changes to its data
whenever we call `save` on the application service. This is so we can
remove data that was previously provided if the associated option is
unchecked.

We add an `onSave` method to the parent class of the form pages, so each
page can implement its own.