The parental dueDate field is present in all scenarios
When an application is submitted for the first time and the user answers ‘Yes’, nunjucks will create a new field called ‘dueDate’, this field is then stored in the `application.data`.

When the answer is ‘No’ the field doesn’t exist on a new application or when the page is presented, introducing a variance that made it difficult to add new unit tests. When creating tests for a person who is __not__ pregnant we tried to provide empty year, month and day fields. We could do this and check for each field to be undefined before parsing it to a date but that felt a little clunky.

A real application had the following set in application data when the answer was ‘yes’, notice the ‘dueDate’:
```
{"isPregnant":"yes","dueDate-month":"3","dueDate-year":"2024","dueDate-day":"27","dueDate":"2024-03-27"}
```

A real application had the following set in application data when the answer was ‘no’, notice the missing ‘dueDate’:
```
{"isPregnant":"no","dueDate-month":"","dueDate-year":"","dueDate-day":""}
```

A consequence of this change is that now `dueDate` will always exist as a key within `application.data`. It will either be filled in with a date, if pregnant, or it will be undefined in the same way `dueDate-year`, `dueDate-month` and `dueDate-day` are.

Co-authored-by: Patrick Fleming <patrick.fleming@dwx.com>