Allow full error trace in Nunjucks
When a global function in a template throws an error, we don’t see the
full trace of the error, which makes diagnosing issues really hard.

This makes use of the [undocumented dev argument](mozilla/nunjucks#1430)
which allows us to see the cause of the error in our logs and in Sentry.

[Copied from CAS1.](https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/490/commits/8eed08b4e8698b0ad16bc285f731bc8b8f0cd3b2)