add '/info' endpoint
This adds an `/info` endpoint for CAS2, which is required in order to
share the prison codes for private beta roll out, as required to have a
tile on DPS[1].

The `activeAgencies` list will in the near future be set as a list of
the prison codes we're rolling out to. An empty list is the preferred
state for a service that is not yet rolled out.

When we are ready to launch private beta, we can update this list with
the codes required, and DPS will be able to pick the information up
within an hour.

When looking at the hmpps-typescript-template repo[2], this `/info`
endpoint also has a productId, as well as info about git and version. As
this is a public repo, I'm not going to add any information except the
`activeAgencies` required, until we need it.
The productId is used by MoJ to compile a service catalogue, which can
be found here https://developer-portal.hmpps.service.justice.gov.uk/products

The template has the route setup in `setUpHealthCheck` but I thought it
made more sense to have this in a separate file, as this is not related
to the `/health` endpoint.

[1]https://dsdmoj.atlassian.net/wiki/spaces/ALIGN/pages/4480959439/Adding+changing+services+on+the+common+header+footer+and+the+homepage#Establishment-specific-rollouts
[2]https://github.com/ministryofjustice/hmpps-template-typescript/blob/main/server/middleware/setUpHealthChecks.ts#L24