Install Playwright
It’s important to note that e2e tests are required to pass for deploying the frontend to preprod and prod.

We found that it can be difficult to coordinate the frontend changes (a pull request in 1 repo) alongside the changes to the e2e tests (a pull request in another repo).

Both must be worked on and tested locally from different repos. We also have double the admin of proposing and reverting two pull requests, rather than just one.

We’d like to move these tests from a separate repo to the same repo.