Update comment for viewing CYA page
When this code was first introduced, it allowed the form to be submitted
without having completed all the tasks. After changes to the CYA page
and how we submit the form, this code now allows us to view the CYA page
without completing all tasks.

In future, the check your answers page will be disabled until all tasks
are complete, but until that's implemented, we'll keep this logic in.