Create submittedApplicationDetails partial and update assess overview template
We create a new partial for rendering the application details, which we will use in the next commit for our new referrer overview template. The assess template is updated to use this partial.

We also move the timeline partial, as it will now be used by both the apply and assess sides of the service.