Introduce status update details controller
Assessors require the ability to choose a secondary status when updating
the status of the application. This will allow for more specific
statuses.

This commit adds the controller which will be responsible for rendering
the new page.

As we are relying on a URL param, we also check it against the list of
accepted status names and redirect to the initial status page if it is
invalid.