Introduce status update details view
Assessors require the ability to choose a secondary status when updating
the status of the application. This will allow for more specific
statuses.

This view will be rendered when the initial status chosen has associated
statusDetails. If the status is "moreInfoRequested" we render
checkboxes, and in all other cases, radio inputs.