Redirect to details page from status update controller
Assessors require the ability to choose a secondary status when updating
the status of the application. This will allow for more specific
statuses.

Some statuses will update the application directly and some will
redirect to another page to choose the secondary statuses. We decide
this by checking whether the chosen status contains status details in
its reference data.