e2e testing for admin/assessor journey
now that pagination has been introduced to the
listing page for submitted applications, we need
to be less specific about the application that the
admin or assessor views.

Previously we looked for the name of the applicant
created by the referrer in `01_apply.spec.ts`, but
because we don't know how many applications have
been created on dev, we don't know what page that
application will appear on.

We know that there will always be at least 1
application submitted, because of the seeded data,
so this changes the test to grab the first link,
and then checks the first two expected headings
are there.