Introduce util for rendering status detail options
Assessors require the ability to choose a secondary status when updating
the status of the application. This will allow for more specific
statuses.

If an assessor updates the application to a status that contains
`statusDetails`, they will be redirected to a status detail page which
will render either checkboxes or radios for each sub-status.