Introduce utils for kebab and camel case
Assessors require the ability to choose a secondary status when updating
the status of the application. This will allow for more specific
statuses.

In order to persist the the status name between the two status pages
we'll be reusing the status name from the URL params. As the URL will be
kebab case and the usage requires it to be camelCase, we'll make use of
these helpers.