Introduce application current status partial
As we will be introducing a new status detail page, we'll want to reuse
the status header which exists on the original page.

We create a partial for this and update the styling to stop the status
from breaking onto two lines when the status is very long.

We also make use of this partial in the submitted application details
partial.