Set back button to overview page if status details is referer
Assessors require the ability to choose a secondary status when updating
the status of the application. This will allow for more specific
statuses.

As a result of optionally rendering another page when a status is
selected, it's possible to get stuck in a loop when using the back
button on the initial status update page due to it setting the URL to
the referer.

To avoid this, we set the back button to take the user back to the
overview page if the referer is the new status details page.