Introduce util for retrieving status details questions
Assessors require the ability to choose a secondary status when
updating the status of the application.

This util returns the question associated with the status detail name.
If none is found, we return an empty string.

If in future the status details being returned from the API changes, our
integration tests will fail as we will also make use of this util in our
test pages.