Introduce formatCommaToLinebreak util
We are expanding the timeline component to show different types of
timeline events: Notes, status updates and when an application has been
created.

For status updates we want to be able to format the body of these events
when it contains a list. This commit introduces a util for this
functionality.