Use application.timelineEvents
We are expanding the timeline component to show different types of
timeline events: Notes, status updates and when an application has been
created.

This commit updates the util for formatting the timeline events to use
`application.timelineEvents` which will contain all of the different
types of events.

We add a couple of conditionals:

* Format status updates which contain lists
* If the feature flag is disabled do not show notes