Error when previous conviction answer doesn't match
As we have changed the options to the previous convictions question, the
answer keys have changed but if the question was answered previously,
this data will still exist.

This means in progress applications won't change the status of the task
to incomplete and force the referrer to update their answer[1].

Here we check that the answer contained in the body always matches the
expected answers. We make use of an enum to keep these values
consistent.

With this in place, we ensure the correct status is applied to the
task[2].

[1]
https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-ui/blob/cas2-227-changes-to-previous-convictions-question/server/utils/checkYourAnswersUtils.ts#L86

[2]
https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-ui/blob/3ec1bb50db30d8908169b7c495eed73bf34d0266/server/form-pages/utils/getTaskStatus.ts#L39