Extend E2E tests to check note added appears on application timeline
Because of the way the tests are written, we will likely end up adding notes to the same applications multiple times through the E2E tests. So, we have a double safeguard of checking that the note we have added is first in the list, and also that it is unique.