Add API assessment paths
We need to include identifying information about each submitted
application so that it can be included in the email notifications that
are sent to the assessor inbox. Therefore, we are adding a new page for
assessors to enter this information.

In this commit we add the API endpoint for assessments.