Add link to assessor details page
We need to include identifying information about each submitted
application so that it can be included in the email notifications that
are sent to the assessor inbox. Therefore, we are adding a new page for
assessors to enter this information.

In this commit we add a link to add assessment details. We change the
copy to "change" if there are already assessment details present.