Introduce assessment utils
We need to include identifying information about each submitted
application so that it can be included in the email notifications that
are sent to the assessor inbox. Therefore, we are adding a new page for
assessors to enter this information.

We create a util to determine whether assessment details have already
been added. We'll use this to change the copy presented to the assessor.